etl/requirements.txt
requests==2.31.0
supabase==2.5.3
python-dateutil==2.9.0
openpyxl==3.1.2
python-dotenv==1.0.1
